+++ src/components/explore/Tabs.jsx
@@ -17,6 +17,7 @@ const tabs = [
     { id: 'humidity', label: i18n.t('Humidity') },
     { id: 'heat', label: i18n.t('Heat') },
     { id: 'climatechange', label: i18n.t('Climate change') },
+    { id: 'vegetation', label: i18n.t('Vegetation') },
 ]
 
 const Tabs = () => {